requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.63
langchain-openai==0.3.19
fastapi
uvicorn
jinja2
python-multipart
mem0ai
requests